Unify slides, use original slide if the single-votes format won't work
