+++ client/src/app/site/pages/meetings/modules/projector/modules/slides/components/poll-slide/components/poll-slide.component.html
@@ -1,28 +1,111 @@
 @if (data?.data) {
-    <div class="slidetitle" [ngStyle]="titleDivStyles">
+    <div class="slidetitle" [ngClass]="{ 'single-votes-slidetitle': useSingleVotes }" [ngStyle]="titleDivStyles">
         <ng-container>
-            <h1>{{ title }}</h1>
+            <h1 [ngClass]="{ 'ellipsis-overflow': useSingleVotes }">{{ title }}</h1>
             @if (subtitle) {
-                <h2 class="poll-title">{{ subtitle }}</h2>
+                <h2 class="poll-title" [ngClass]="{ 'ellipsis-overflow': useSingleVotes }">{{ subtitle }}</h2>
             }
         </ng-container>
     </div>
+    @if (useSingleVotes) {
+        <div class="poll-bar-wrapper">
+            <div class="poll-bar">
+                @if (polldata?.options[0].yes) {
+                    <div class="bg-yes" [ngStyle]="{ flex: polldata.options[0].yes + ' 0 auto' }"></div>
+                }
+                @if (polldata?.options[0].no) {
+                    <div class="bg-no" [ngStyle]="{ flex: polldata.options[0].no + ' 0 auto' }"></div>
+                }
+                @if (polldata?.options[0].abstain) {
+                    <div class="bg-abstain" [ngStyle]="{ flex: polldata.options[0].abstain + ' 0 auto' }"></div>
+                }
+            </div>
+            <div class="data-area">
+                <div class="general-data">
+                    <div>
+                        {{ 'Valid votes' | translate }} {{ polldata.votesvalid }}
+                        {{ polldata.votesvalid | pollPercentBase: polldata }}
+                    </div>
+                    <div>
+                        {{ 'Entitled users' | translate }} {{ polldata.entitled_users_at_stop.length }}
+                        @if (polldata.onehundred_percent_base === 'entitled') {
+                            {{ polldata.entitled_users_at_stop.length | pollPercentBase: polldata }}
+                        }
+                    </div>
+                </div>
+                <os-icon-container icon="check_circle" iconClass="yes" size="large">
+                    <span class="yes">
+                        {{ 'Yes' | translate }} {{ polldata.options[0]?.yes ?? 0 }}
+                        {{ polldata.options[0]?.yes ?? 0 | pollPercentBase: polldata }}
+                    </span>
+                </os-icon-container>
+                <os-icon-container icon="cancel" iconClass="no" size="large">
+                    <span class="no">
+                        {{ 'No' | translate }} {{ polldata.options[0]?.no ?? 0 }}
+                        {{ polldata.options[0]?.no ?? 0 | pollPercentBase: polldata }}
+                    </span>
+                </os-icon-container>
+                <os-icon-container icon="circle" iconClass="abstain" size="large">
+                    <span class="abstain">
+                        {{ 'Abstain' | translate }} {{ polldata.options[0]?.abstain ?? 0 }}
+                        {{ polldata.options[0]?.abstain ?? 0 | pollPercentBase: polldata }}
+                    </span>
+                </os-icon-container>
+            </div>
+        </div>
+    }
     @if (polldata) {
         <div class="content-wrapper">
             <div [ngStyle]="textDivStyles">
-                @if (pollContentObjectType === PollContentObjectType.Motion) {
-                    <div class="motion-poll-slide">
-                        <os-motion-poll-detail-content
-                            iconSize="gigantic"
-                            [poll]="polldata"
-                        ></os-motion-poll-detail-content>
+                @if (useSingleVotes) {
+                    <div class="poll-result-area">
+                        <div class="votes-area">
+                            @for (column of userVotesFormatted; track column) {
+                                <div class="votes-column" [ngStyle]="columnStyle">
+                                    @for (user of column; track user[0]) {
+                                        <div class="vote-entry">
+                                            @if (user[1] === 'Y') {
+                                                <os-icon-container icon="check_circle" iconClass="yes" [noWrap]="true">
+                                                    {{ user[0] }}
+                                                </os-icon-container>
+                                            } @else if (user[1] === 'N') {
+                                                <os-icon-container icon="cancel" iconClass="no" [noWrap]="true">
+                                                    {{ user[0] }}
+                                                </os-icon-container>
+                                            } @else if (user[1] === 'A') {
+                                                <os-icon-container icon="circle" iconClass="abstain" [noWrap]="true">
+                                                    {{ user[0] }}
+                                                </os-icon-container>
+                                            } @else {
+                                                <os-icon-container
+                                                    icon="radio_button_unchecked"
+                                                    iconClass="no-content"
+                                                    [noWrap]="true"
+                                                >
+                                                    {{ user[0] }}
+                                                </os-icon-container>
+                                            }
+                                        </div>
+                                    }
+                                </div>
+                            }
+                        </div>
                     </div>
-                }
-                @if (pollContentObjectType === PollContentObjectType.Assignment) {
-                    <os-assignment-poll-detail-content [poll]="polldata"></os-assignment-poll-detail-content>
-                }
-                @if (pollContentObjectType === PollContentObjectType.Topic) {
-                    <os-topic-poll-detail-content [poll]="polldata"></os-topic-poll-detail-content>
+                } @else {
+                    @if (pollContentObjectType === PollContentObjectType.Motion) {
+                        <div class="motion-poll-slide">
+                            <os-motion-poll-detail-content
+                                iconSize="gigantic"
+                                [poll]="polldata"
+                            ></os-motion-poll-detail-content>
+                        </div>
+                    }
+                    @if (pollContentObjectType === PollContentObjectType.Assignment) {
+                        <os-assignment-poll-detail-content [poll]="polldata"></os-assignment-poll-detail-content>
+                    }
+                    @if (pollContentObjectType === PollContentObjectType.Topic) {
+                        <os-topic-poll-detail-content [poll]="polldata"></os-topic-poll-detail-content>
+                    }
                 }
             </div>
         </div>
