+++ client/src/app/site/pages/meetings/modules/projector/modules/slides/components/poll-slide/components/poll-slide.component.ts
@@ -1,19 +1,29 @@
-import { Component, Input } from '@angular/core';
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component, Input, OnDestroy } from '@angular/core';
 import { _ } from '@ngx-translate/core';
-import { of } from 'rxjs';
+import { of, Subscription } from 'rxjs';
 import { OptionData, OptionTitle, PollData } from 'src/app/domain/models/poll/generic-poll';
 import { PollClassType, PollState } from 'src/app/domain/models/poll/poll-constants';
 import { Projector } from 'src/app/domain/models/projector/projector';
+import { User } from 'src/app/domain/models/users/user';
 import { collectionFromFqid } from 'src/app/infrastructure/utils/transform-functions';
 import { UnknownUserLabel } from 'src/app/site/pages/meetings/pages/assignments/modules/assignment-poll/services/assignment-poll.service';
 import { ViewProjector } from 'src/app/site/pages/meetings/pages/projectors';
 import { SlideData } from 'src/app/site/pages/meetings/pages/projectors/definitions';
+import { MeetingSettingsService } from 'src/app/site/pages/meetings/services/meeting-settings.service';
 import { CollectionMapperService } from 'src/app/site/services/collection-mapper.service';
 
 import { BaseScaleScrollSlideComponent } from '../../../base/base-scale-scroll-slide-component';
 import { BaseSlideComponent } from '../../../base/base-slide-component';
 import { modifyAgendaItemNumber } from '../../../definitions';
-import { PollSlideData, PollSlideDataFields, SlidePollOption, SlidePollOptionFields } from '../poll-slide-data';
+import {
+    PollSlideData,
+    PollSlideDataFields,
+    PollSlideEntitledUsersEntry,
+    SlidePollOption,
+    SlidePollOptionFields,
+    SlidePollUser,
+    SlidePollVote
+} from '../poll-slide-data';
 
 export enum PollContentObjectType {
     Standalone = `standalone`,
@@ -22,14 +32,22 @@ export enum PollContentObjectType {
     Topic = `topic`
 }
 
+type VoteResult = `Y` | `N` | `A` | `X`;
+const ENTRY_HEIGHT = 30;
+const TITLE_HEIGHT = 55;
+const POLL_BAR_HEIGHT = 91;
+const HEADER_FOOTER_HEIGHT = 125;
+const NO_HEADER_TOP_MARGIN = 40;
+
 @Component({
     selector: `os-poll-slide`,
     templateUrl: `./poll-slide.component.html`,
-    styleUrls: [`./poll-slide.component.scss`]
+    styleUrls: [`./poll-slide.component.scss`],
+    changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class PollSlideComponent
     extends BaseSlideComponent<PollSlideData>
-    implements BaseScaleScrollSlideComponent<PollSlideData>
+    implements BaseScaleScrollSlideComponent<PollSlideData>, OnDestroy
 {
     public PollState = PollState;
     public PollContentObjectType = PollContentObjectType;
@@ -76,7 +94,7 @@ export class PollSlideComponent
     @Input()
     public set scroll(scroll: number) {
         this._scroll = scroll ?? 0;
-        scroll *= -100;
+        scroll *= -25;
         this.textDivStyles[`margin-top`] = `${scroll}px`;
         this.onScroll();
     }
@@ -101,8 +119,62 @@ export class PollSlideComponent
         return this._scale;
     }
 
-    public constructor(private collectionMapperService: CollectionMapperService) {
+    // SINGLE VOTES
+
+    public get useSingleVotes(): boolean {
+        return (
+            this._isSingleVotes &&
+            !this._invalidSingleVotesData &&
+            this.pollContentObjectType === PollContentObjectType.Motion &&
+            !this.data.data.is_pseudoanonymized
+        );
+    }
+
+    private _isSingleVotes: boolean = false;
+
+    private _invalidSingleVotesData = false;
+
+    public gridStyle: { [key: string]: any } = {};
+
+    public columnStyle: { [key: string]: any } = {};
+
+    public bufferUp: number;
+    public userVotesFormatted: [string, VoteResult][][];
+
+    private _maxColumns = 6;
+    private _orderBy: keyof User = `last_name`;
+    private _userVotes: [string, VoteResult][] = [];
+
+    private _votes: {
+        [key: string]: SlidePollVote;
+    } = {};
+
+    private _entitledUsers: {
+        [key: string]: PollSlideEntitledUsersEntry;
+    } = {};
+
+    private _meetingSettingsSubscriptions: Subscription[];
+
+    public constructor(
+        private collectionMapperService: CollectionMapperService,
+        private cd: ChangeDetectorRef,
+        private meetingSettings: MeetingSettingsService
+    ) {
         super();
+        this._meetingSettingsSubscriptions = [
+            this.meetingSettings.get(`motion_poll_projection_max_columns`).subscribe(max => {
+                this._maxColumns = max;
+                this.formatUserVotes();
+            }),
+            this.meetingSettings.get(`motion_poll_projection_name_order_first`).subscribe(orderBy => {
+                this._orderBy = orderBy;
+                this.calculateUserVotes();
+            })
+        ];
+    }
+
+    public ngOnDestroy(): void {
+        this._meetingSettingsSubscriptions.forEach(sub => sub.unsubscribe());
     }
 
     protected onProjectorFormatChange(projector: ViewProjector): void {
@@ -111,13 +183,35 @@ export class PollSlideComponent
         } else {
             this.titleDivStyles[`margin-top`] = `0`;
         }
+        this.formatUserVotes();
     }
 
-    protected onScroll(): void {}
+    protected onScroll(): void {
+        this.cd.markForCheck();
+    }
 
-    protected onScale(): void {}
+    protected onScale(): void {
+        this.formatUserVotes();
+    }
 
     protected override setData(value: SlideData<PollSlideData>): void {
+        const poll = value.data;
+        if (value.type == `poll_single_votes` && poll && poll.options.length === 1) {
+            this._isSingleVotes = true;
+            this._votes =
+                poll.options[0].votes?.mapToObject(vote => ({
+                    [vote.user_id]: vote
+                })) || {};
+            this._entitledUsers =
+                value.data.entitled_users_at_stop?.mapToObject(user => ({
+                    [user.user_merged_into_id ?? user.user_id]: user
+                })) || {};
+            this.calculateUserVotes();
+            this._invalidSingleVotesData = false;
+        } else {
+            this._invalidSingleVotesData = true;
+        }
+
         super.setData(value);
 
         // Convert every decimal(string) to a float
@@ -160,6 +254,7 @@ export class PollSlideComponent
             this.title = value.data.title;
             this.subtitle = null;
         }
+        this.cd.markForCheck();
     }
 
     private createPollData(data: PollSlideData): PollData {
@@ -217,5 +312,95 @@ export class PollSlideComponent
             weight,
             ...(data.content_object ? { entries_amount: data.content_object[`entries_amount`] } : {})
         };
+    }
+
+    private calculateUserVotes(): void {
+        if (this._isSingleVotes) {
+            this._userVotes = Array.from(new Set([...Object.keys(this._votes), ...Object.keys(this._entitledUsers)]))
+                .map(id => [
+                    ...(this.getNameAndSortValue(
+                        this._entitledUsers[id]?.user || this._votes[id]?.user,
+                        this._orderBy
+                    ) || [`User`, `${id}`]),
+                    this._votes[id] ? this._votes[id].value : `X`
+                ])
+                .sort((a, b) => a[1].localeCompare(b[1]))
+                .map(([user, _, vote]) => [user, vote as VoteResult]);
+            this.formatUserVotes();
+        }
+    }
+
+    private getNameAndSortValue(user: SlidePollUser, by: keyof User): [string, string] {
+        if (user == undefined) {
+            return null;
+        }
+
+        const first = `${user.title ?? ``} ${user.first_name ?? ``}`.trim();
+        if (first && user.last_name) {
+            if (by === `last_name`) {
+                return [`${user.last_name}, ${first}`, `${user.last_name}, ${user.first_name}`];
+            }
+            return [
+                `${first} ${user.last_name}`,
+                user.first_name ? `${user.first_name}, ${user.last_name}` : user.last_name
+            ];
+        }
+        if (first) {
+            if (by === `last_name`) {
+                return [first, `${user.first_name}`];
+            }
+            return [first, user.first_name];
+        }
+        if (user.last_name) {
+            if (by === `last_name`) {
+                return [user.last_name, user.last_name];
+            }
+            return [user.last_name, `${user.last_name}`];
+        }
+        return null;
+    }
+
+    private formatUserVotes(): void {
+        if (!this.projector || !this._userVotes || !this._isSingleVotes) {
+            return;
+        }
+        const [columns, rows] = this.newGetUserVotesFormat();
+        this.newCalculateFormattedUserVotes(columns, rows);
+        `1fr `.repeat(columns).trim();
+        this.gridStyle = {
+            [`grid-template-columns`]: `1fr `.repeat(columns).trim()
+        };
+        this.columnStyle = {
+            [`width`]: `${Math.floor(((this.projector?.width ?? 100) - 100) / columns - 5)}px`
+        };
+        this.cd.markForCheck();
+    }
+
+    private newGetUserVotesFormat(): [number, number] {
+        const visibleHeight =
+            (this.projector.height -
+                TITLE_HEIGHT -
+                POLL_BAR_HEIGHT -
+                (this.projector.show_header_footer ? HEADER_FOOTER_HEIGHT : NO_HEADER_TOP_MARGIN)) /
+            this._actualScale;
+        const visibleRows = Math.floor(visibleHeight / ENTRY_HEIGHT);
+        const columns = Math.min(Math.ceil(this._userVotes.length / visibleRows), this._maxColumns);
+        return [columns, Math.max(visibleRows, Math.ceil(this._userVotes.length / columns))];
+    }
+
+    private newCalculateFormattedUserVotes(columns: number, rows: number): void {
+        let nextIndex = 0;
+        const votesFormatted: [string, VoteResult][][] = [];
+        for (let i = 0; i < columns; i++) {
+            const [stop, untilIndex] = this.newCalcStopAndActualUntilIndex(nextIndex + rows);
+            votesFormatted.push(this._userVotes.slice(nextIndex, untilIndex));
+            if (stop) break;
+            nextIndex = untilIndex;
+        }
+        this.userVotesFormatted = votesFormatted;
+    }
+
+    private newCalcStopAndActualUntilIndex(untilIndex: number): [boolean, number] {
+        return [untilIndex >= this._userVotes.length, Math.min(untilIndex, this._userVotes.length)];
     }
 }
