+++ client/src/app/site/pages/meetings/modules/projector/modules/slides/components/poll-slide/poll-slide-data.ts
@@ -63,6 +63,7 @@ export interface PollSlideData {
     votesinvalid: number;
     votescast: number;
     global_option: SlidePollOption;
+    is_pseudoanonymized: boolean;
 }
 
 export const PollSlideDataFields: (keyof PollSlideData)[] = [`votesvalid`, `votesinvalid`, `votescast`];
